DEPLOY_ZKL update channel topic handling to use double for time
svn path=/trunk/other/; revision=11623

+++ Shared/LobbyClient/TasClient.cs
@@ -371,7 +371,7 @@ namespace LobbyClient
 
         public static DateTime ConvertMilisecondTime(string arg)
         {
-            return (new DateTime(1970, 1, 1, 0, 0, 0)).AddMilliseconds(long.Parse(arg));
+            return (new DateTime(1970, 1, 1, 0, 0, 0)).AddMilliseconds(double.Parse(arg));
         }
 
         public void DisableUnits(params string[] units)
